Reworking the AP-M1-Mini description.
Introducing a sane concept of elementary building blocks and internal
wiring (partlists in entities). Also introducing "is-a" rules for
inheritance of types (which are however not yet implemented).

What is not so sane is that I don't always follow the rule "describe
input only", for instance entity partlists describe output as well as
wired parts still do.

In theory, this level of description should allow to model the whole
wiring of the computer.