DDA: Various smallish improvements
Mostly on the C++ runner, which can now print progress bars and is
faster for reading in large amounts of data.

Furthermore, some old and new tests. Haven't checked which one run. I
think I broke the API in the last 2 days anyway and all checks have to
be checked manually :)